+++ python/xir/interfaces/__init__.py
@@ -1,1 +1,0 @@
-from .strawberryfields_io import to_program, to_xir